Move route protectors and update directories
