+++ src/components/RouteProtectors/IsPrivate.jsx
@@ -2,7 +2,7 @@
 
 import { useContext } from 'react';
 import { Navigate } from 'react-router-dom';
-import { AuthContext } from '../context/auth.context';
+import { AuthContext } from '../../context/auth.context';
 
 function IsPrivate({ children }) {
 	const { isLoggedIn, isLoading } = useContext(AuthContext);
